add stall-length to `driveCircuit`
This allows a test to control the "downtime" between two
transactions.

The `stallCircuit` does not yet work with "zero-downtime"
transactions. It also only used to test the compliance of the
Wishbone master, which in `wishbonePropWithModel` was only ever
`driveCircuit`. This was removed for now because it doesn't add
trust to the circuit-under-test.

This commit also adds more compact `Show` and `ShowX` instances
for Wishbone{S2M|M2S} values.