Support GHC 9.8 and 9.10
GHC 9.8 and 9.10 give a lot of Haddock warnings. We used to check the
Haddock for all GHC versions, but this is overkill. Just check it on GHC
9.6 only.

We used to first invoke `cabal build` and then
`cabal build --enable-documentation`. But this rebuilds all
dependencies, effectively doubling running time.

The `allow-newer` on `circuit-notation` is superfluous.

Further details: Haddock for GHC 9.8 and up generate more information
about type family instances. However, this also leads to "could not find
link destinations for" warnings. Given an `instance C A`, it would seem
that when it lists the instance under the Haddock of the definition of
`C`, it tries to create a link to the listing of the instance under the
Haddock of the definition of `A`, and fails somehow. I suspect a bug in
Haddock rather than something we can fix.