Fix use of deprecated OPTIONS pragma
Additionally, the Haddock module description seemed to refer to the
`-fplugin` option rather than actually describing the module, so the
comment is moved and made a regular comment instead of a Haddock block.
For consistency with the rest of the code base, line comments are used
instead of block comments.