Remove strict tuples in AXI protocol
This is a prerequisite to adding `BitPack` instances to all AXI channel signals.

Martijn on why we can switch away from strict tuples: "I started out
clash-protocols with many strict data types out of a fear of space leaks. But
this was never really well defended in the first place, so IMO we should just do
what the rest of the Haskell ecosystem does by default: laziness."

Co-authored-by: Martijn Bastiaan <martijn@qbaylogic.com>