Optimize ready signaling
Many streaming components currently insist that the READY signal takes
on a particular value when VALID is driven low. This introduces a
combinational path between IN_VALID and OUT_READY on the same streaming
interface, which is unnecessary. Driving READY as X (don't care) when
VALID is driven low removes these paths, likely resulting in better
timing (and maybe a tiny bit of logic saved).