WIP, `propWithModel` return `PropertyT IO ()` instead of `Property`
Currently the user of `propWithModel` can not generate any aspects of the arguments to this function. The cause is the fact that `propWithModel` uses the `property` function from `Hedgehog`.

If we don't call that function, but return the monad instead, the user can generate values for the arguments to `propWithModel`, such as configuration parameters of their circuit.