Rework Nix systems (#148)
Removed all 'old-style' Nix infrastructure in favor of nix flakes.

This makes all the Nix logic easier to understand whilst also making it
easier to include this within other projects.

It also exposes circuit-notation as a package, as you often want to use
this if you use, well, circuit notation. This makes it easy to include
it as a nix dependency.